backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.9.0
django-redis==5.4.0
djangorestframework==3.16.1
drf-spectacular==0.27.2
djangorestframework_simplejwt==5.5.1
django-storages==1.14.4
boto3==1.35.84
coverage==7.6.10
PyJWT==2.10.1
python-dotenv==1.2.1
Pillow==10.4.0
sqlparse==0.5.5
tzdata==2025.3
gunicorn==23.0.0
whitenoise[brotli]==6.8.2
psycopg[binary]==3.2.3
redis==5.2.1
sentry-sdk==2.20.0
